fix(modal): change the mouse event to dismiss a modal on backdrop click (#5326)
* fix(modal): change the mouse event to dismiss a modal on backdrop click

Because of the use of a 'click' event to trigger the dismissal of a modal, the target of the event doesn't take into account an eventual movement of the mouse before releasing click (for ex. selection of text), which can lead to unwanted dismissal of the modal.

Closes: #5264

* fix(modal): change the mouse event to dismiss a modal from a directive

Because of the use of a 'click' event to trigger the dismissal of a modal, the target of the event doesn't take into account an eventual movement of the mouse before releasing click (for ex. selection of text), which can lead to unwanted dismissal of the modal.