fix(typeahead): add variable which was missed in MR #5547
adding variable `list` which was added in host Object as `[attr.aria-autocomplete]: 'list'` but was not defined.
Modules/Libraries which uses AoT compilation for their build gets fatal error due to this.

this will fix let AoT builds to work properly which was introduced in MR https://github.com/valor-software/ngx-bootstrap/pull/5547